requirements.txt
openai
gradio
mcp[cli]
requests
python-dotenv
uvicorn
